fix10

+++ app/api/games/delete.ts
@@ -1,20 +1,25 @@
 import { NextApiRequest, NextApiResponse } from "next";
-import prisma from "../../../prisma/client";
+import prisma from "../../../utils/db";
 
 export default async function handler(req: NextApiRequest, res: NextApiResponse) {
   if (req.method === "DELETE") {
-    const { id } = req.query;
-
     try {
-      const deletedGame = await prisma.game.delete({
-        where: { id: Number(id) },
+      const { id } = req.query;
+
+      if (!id) {
+        return res.status(400).json({ message: "ID is required." });
+      }
+
+      await prisma.game.delete({
+        where: { id: parseInt(id as string) },
       });
-      res.status(200).json(deletedGame);
+
+      return res.status(200).json({ message: "Game deleted successfully." });
     } catch (error) {
       console.error("Error deleting game:", error);
-      res.status(500).json({ error: "Failed to delete game." });
+      return res.status(500).json({ message: "Internal server error." });
     }
   } else {
-    res.status(405).json({ message: "Method not allowed" });
+    return res.status(405).json({ message: "Method not allowed." });
   }
 }
